@@ -22,15 +22,19 @@ class User extends Authenticatable
         'password',
     ];
 
+    protected $visible = [
+        'name',
+    ];
+
     /**
      * The attributes that should be hidden for arrays.
      *
      * @var array
-     */
-    protected $hidden = [
-        'password',
-        'remember_token',
-    ];
+     protected $hidden = [
+         'password',
+         'remember_token',
+        ];
+        */
 
     /**
      * The attributes that should be cast to native types.
@@ -40,4 +44,9 @@ class User extends Authenticatable
     protected $casts = [
         'email_verified_at' => 'datetime',
     ];
+
+    public function photos()
+    {
+        return $this->hasMany('App\Models\Photo');
+    }
 }
