@@ -34,7 +34,7 @@ return [
     */
 
     'connections' => [
-
+        
         'sqlite' => [
             'driver' => 'sqlite',
             'url' => env('DATABASE_URL'),
