@@ -13,8 +13,9 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        $this->call(CitySeeder::class);
         $this->call(RegionSeeder::class);
+        $this->call(PrefectureSeeder::class);
+        $this->call(CitySeeder::class);
         // \App\Models\User::factory(10)->create();
     }
 }
