@@ -14,13 +14,13 @@
         </div>
     </div>
     <div class="w-50">
-        <img src="{{ cityImage($cities[0]->slug) }}" alt="city" class="w-100">
     </div>
 </div>
 <div class="introduction-section">
     <div class="container">
 
         <div class="cities text-center">
+            {{--
             @foreach ($cities as $city)
                 <div class="city">
                     <a href="{{ route('city.show', $city->slug) }}"><img src="{{ cityImage($city->slug) }}" alt="city"></a>
@@ -28,10 +28,7 @@
                     <div class="city-price"></div>
                 </div>
             @endforeach
-        </div>
-        
-        <div class="text-center button-container">
-            <a href="{{ route('city.index') }}" class="button">View more products</a>
+            --}}
         </div>
     </div> <!-- end container -->
 </div>
