@@ -11,5 +11,11 @@ const mix = require('laravel-mix');
  |
  */
 
-mix.js('resources/js/app.js', 'public/js')
+ /*
+mix.b.js('resources/js/app.js', 'public/js')
     .sass('resources/sass/app.scss', 'public/css');
+*/
+mix.browserSync('rehome.test')
+  .js('resources/js/app.js', 'public/js')
+  .sass('resources/sass/app.scss', 'public/css')
+  .version();
